fix (fix/meeyoung/#14) 쇼핑몰 목록 페이지 컴포넌트로 분리
